requirements.txt
Django==1.11.5
mysqlclient==1.3.12
django-mysql==2.2.
